Conflicts:
    - kani-compiler/src/kani_middle/stubbing/annotations.rs
    - kani-compiler/src/session.rs (logger initialization issue)

+++ kani-driver/src/call_cargo.rs
@@ -2,12 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 use crate::args::KaniArgs;
-use crate::session::{KaniSession, ReachabilityMode};
+use crate::call_single_file::to_rustc_arg;
+use crate::project::Artifact;
+use crate::session::KaniSession;
+use crate::util;
 use anyhow::{bail, Context, Result};
-use cargo_metadata::{Metadata, MetadataCommand, Package};
-use std::ffi::OsString;
-use std::fs;
-use std::path::{Path, PathBuf};
+use cargo_metadata::diagnostic::{Diagnostic, DiagnosticLevel};
+use cargo_metadata::{Message, Metadata, MetadataCommand, Package, Target};
+use kani_metadata::{ArtifactType, CompilerArtifactStub};
+use std::ffi::{OsStr, OsString};
+use std::fmt::{self, Display};
+use std::fs::{self, File};
+use std::io::BufReader;
+use std::path::PathBuf;
 use std::process::Command;
 use tracing::{debug, trace};
 
@@ -26,19 +33,17 @@ pub struct CargoOutputs {
     /// The directory where compiler outputs should be directed.
     /// Usually 'target/BUILD_TRIPLE/debug/deps/'
     pub outdir: PathBuf,
-    /// The collection of *.symtab.json files written.
-    pub symtabs: Vec<PathBuf>,
-    /// The location of vtable restrictions files (a directory of *.restrictions.json)
-    pub restrictions: Option<PathBuf>,
     /// The kani-metadata.json files written by kani-compiler.
-    pub metadata: Vec<PathBuf>,
+    pub metadata: Vec<Artifact>,
     /// Recording the cargo metadata from the build
     pub cargo_metadata: Metadata,
+    /// For build `keep_going` mode, we collect the targets that we failed to compile.
+    pub failed_targets: Option<Vec<String>>,
 }
 
 impl KaniSession {
     /// Calls `cargo_build` to generate `*.symtab.json` files in `target_dir`
-    pub fn cargo_build(&self) -> Result<CargoOutputs> {
+    pub fn cargo_build(&self, keep_going: bool) -> Result<CargoOutputs> {
         let build_target = env!("TARGET"); // see build.rs
         let metadata = self.cargo_metadata(build_target)?;
         let target_dir = self
@@ -50,14 +55,12 @@ impl KaniSession {
             .join("kani");
         let outdir = target_dir.join(build_target).join("debug/deps");
 
-        // Clean directory before building since we are unable to handle cache today.
-        // TODO: https://github.com/model-checking/kani/issues/1736
-        if target_dir.exists() {
+        if self.args.force_build && target_dir.exists() {
             fs::remove_dir_all(&target_dir)?;
         }
 
-        let mut kani_args = self.kani_specific_flags();
-        let rustc_args = self.kani_rustc_flags();
+        let mut rustc_args = self.kani_rustc_flags();
+        rustc_args.push(to_rustc_arg(self.kani_compiler_flags()).into());
 
         let mut cargo_args: Vec<OsString> = vec!["rustc".into()];
         if let Some(path) = &self.args.cargo.manifest_path {
@@ -81,6 +84,10 @@ impl KaniSession {
         cargo_args.push("--target-dir".into());
         cargo_args.push(target_dir.into());
 
+        // Configuration needed to parse cargo compilation status.
+        cargo_args.push("--message-format".into());
+        cargo_args.push("json-diagnostic-rendered-ansi".into());
+
         if self.args.tests {
             // Use test profile in order to pull dev-dependencies and compile using `--test`.
             // Initially the plan was to use `--tests` but that brings in multiple targets.
@@ -93,36 +100,39 @@ impl KaniSession {
         }
 
         // Arguments that will only be passed to the target package.
-        let mut pkg_args: Vec<OsString> = vec![];
-        match self.reachability_mode() {
-            ReachabilityMode::ProofHarnesses => {
-                pkg_args.extend(["--".into(), "--reachability=harnesses".into()]);
-            }
-            ReachabilityMode::AllPubFns => {
-                pkg_args.extend(["--".into(), "--reachability=pub_fns".into()]);
-            }
-            ReachabilityMode::Tests => {
-                pkg_args.extend(["--".into(), "--reachability=tests".into()]);
-            }
-        }
-
-        // Only joing them at the end. All kani flags must come first.
-        kani_args.extend_from_slice(&rustc_args);
+        let mut pkg_args: Vec<String> = vec![];
+        pkg_args.extend(["--".to_string(), self.reachability_arg()]);
 
         let mut found_target = false;
         let packages = packages_to_verify(&self.args, &metadata);
+        let mut artifacts = vec![];
+        let mut failed_targets = vec![];
         for package in packages {
-            for target in package_targets(&self.args, package) {
+            for verification_target in package_targets(&self.args, package) {
                 let mut cmd = Command::new("cargo");
                 cmd.args(&cargo_args)
                     .args(vec!["-p", &package.name])
-                    .args(&target.to_args())
+                    .args(&verification_target.to_args())
                     .args(&pkg_args)
                     .env("RUSTC", &self.kani_compiler)
-                    .env("RUSTFLAGS", "--kani-flags")
-                    .env("KANIFLAGS", &crate::util::join_osstring(&kani_args, " "));
+                    // Use CARGO_ENCODED_RUSTFLAGS instead of RUSTFLAGS is preferred. See
+                    // https://doc.rust-lang.org/cargo/reference/environment-variables.html
+                    .env("CARGO_ENCODED_RUSTFLAGS", rustc_args.join(OsStr::new("\x1f")))
+                    .env("CARGO_TERM_PROGRESS_WHEN", "never");
 
-                self.run_terminal(cmd)?;
+                match self.run_cargo(cmd, verification_target.target()) {
+                    Err(err) => {
+                        if keep_going {
+                            let target_str = format!("{verification_target}");
+                            util::error(&format!("Failed to compile {target_str}"));
+                            failed_targets.push(target_str);
+                        } else {
+                            return Err(err);
+                        }
+                    }
+                    Ok(Some(artifact)) => artifacts.push(artifact),
+                    Ok(None) => {}
+                }
                 found_target = true;
             }
         }
@@ -132,11 +142,10 @@ impl KaniSession {
         }
 
         Ok(CargoOutputs {
-            outdir: outdir.clone(),
-            symtabs: glob(&outdir.join("*.symtab.json"))?,
-            metadata: glob(&outdir.join("*.kani-metadata.json"))?,
-            restrictions: self.args.restrict_vtable().then_some(outdir),
+            outdir,
+            metadata: artifacts,
             cargo_metadata: metadata,
+            failed_targets: keep_going.then_some(failed_targets),
         })
     }
 
@@ -167,15 +176,89 @@ impl KaniSession {
 
         cmd.exec().context("Failed to get cargo metadata.")
     }
+
+    /// Run cargo and collect any error found.
+    /// We also collect the metadata file generated during compilation if any.
+    fn run_cargo(&self, cargo_cmd: Command, target: &Target) -> Result<Option<Artifact>> {
+        let support_color = atty::is(atty::Stream::Stdout);
+        let mut artifact = None;
+        if let Some(mut cargo_process) = self.run_piped(cargo_cmd)? {
+            let reader = BufReader::new(cargo_process.stdout.take().unwrap());
+            let mut error_count = 0;
+            for message in Message::parse_stream(reader) {
+                let message = message.unwrap();
+                match message {
+                    Message::CompilerMessage(msg) => match msg.message.level {
+                        DiagnosticLevel::FailureNote => {
+                            print_msg(&msg.message, support_color)?;
+                        }
+                        DiagnosticLevel::Error => {
+                            error_count += 1;
+                            print_msg(&msg.message, support_color)?;
+                        }
+                        DiagnosticLevel::Ice => {
+                            print_msg(&msg.message, support_color)?;
+                            let _ = cargo_process.wait();
+                            return Err(anyhow::Error::msg(msg.message).context(format!(
+                                "Failed to compile `{}` due to an internal compiler error.",
+                                msg.target.name
+                            )));
+                        }
+                        _ => {
+                            if !self.args.quiet {
+                                print_msg(&msg.message, support_color)?;
+                            }
+                        }
+                    },
+                    Message::CompilerArtifact(rustc_artifact) => {
+                        if rustc_artifact.target == *target {
+                            debug_assert!(
+                                artifact.is_none(),
+                                "expected only one artifact for `{target:?}`",
+                            );
+                            artifact = Some(rustc_artifact);
+                        }
+                    }
+                    Message::BuildScriptExecuted(_) | Message::BuildFinished(_) => {
+                        // do nothing
+                    }
+                    Message::TextLine(msg) => {
+                        if !self.args.quiet {
+                            println!("{msg}");
+                        }
+                    }
+
+                    // Non-exhaustive enum.
+                    _ => {
+                        if !self.args.quiet {
+                            println!("{message:?}");
+                        }
+                    }
+                }
+            }
+            let status = cargo_process.wait()?;
+            if !status.success() {
+                bail!(
+                    "Failed to execute cargo ({status}). Found {error_count} compilation errors."
+                );
+            }
+        }
+        // We generate kani specific artifacts only for the build target. The build target is
+        // always the last artifact generated in a build, and all the other artifacts are related
+        // to dependencies or build scripts. Hence, we need to invoke `map_kani_artifact` only
+        // for the last compiler artifact.
+        Ok(artifact.and_then(map_kani_artifact))
+    }
 }
 
-/// Given a `path` with glob characters in it (e.g. `*.json`), return a vector of matching files
-fn glob(path: &Path) -> Result<Vec<PathBuf>> {
-    let results = glob::glob(path.to_str().context("Non-UTF-8 path enountered")?)?;
-    // the logic to turn "Iter<Result<T, E>>" into "Result<Vec<T>, E>" doesn't play well
-    // with anyhow, so a type annotation is required
-    let v: core::result::Result<Vec<PathBuf>, glob::GlobError> = results.collect();
-    Ok(v?)
+/// Print the compiler message following the coloring schema.
+fn print_msg(diagnostic: &Diagnostic, use_rendered: bool) -> Result<()> {
+    if use_rendered {
+        print!("{diagnostic}");
+    } else {
+        print!("{}", console::strip_ansi_codes(diagnostic.rendered.as_ref().unwrap()));
+    }
+    Ok(())
 }
 
 /// Extract the packages that should be verified.
@@ -209,20 +292,80 @@ fn packages_to_verify<'b>(args: &KaniArgs, metadata: &'b Metadata) -> Vec<&'b Pa
     packages
 }
 
+/// Extract Kani artifact that might've been generated from a given rustc artifact.
+/// Not every rustc artifact will map to a kani artifact, hence the `Option<>`.
+///
+/// Unfortunately, we cannot always rely on the messages to get the path for the original artifact
+/// that `rustc` produces. So we hack the content of the output path to point to the original
+/// metadata file. See <https://github.com/model-checking/kani/issues/2234> for more details.
+fn map_kani_artifact(rustc_artifact: cargo_metadata::Artifact) -> Option<Artifact> {
+    debug!(?rustc_artifact, "map_kani_artifact");
+    if rustc_artifact.target.is_custom_build() {
+        // We don't verify custom builds.
+        return None;
+    }
+    let result = rustc_artifact.filenames.iter().find_map(|path| {
+        if path.extension() == Some("rmeta") {
+            let file_stem = path.file_stem()?.strip_prefix("lib")?;
+            let parent =
+                path.parent().map(|p| p.as_std_path().to_path_buf()).unwrap_or(PathBuf::new());
+            let mut meta_path = parent.join(file_stem);
+            meta_path.set_extension(ArtifactType::Metadata);
+            trace!(rmeta=?path, kani_meta=?meta_path.display(), "map_kani_artifact");
+
+            // This will check if the file exists and we just skip if it doesn't.
+            Artifact::try_new(&meta_path, ArtifactType::Metadata).ok()
+        } else if path.extension() == Some("rlib") {
+            // We skip `rlib` files since we should also generate a `rmeta`.
+            trace!(rlib=?path, "map_kani_artifact");
+            None
+        } else {
+            // For all the other cases we write the path of the metadata into the output file.
+            // The compiler should always write a valid stub into the artifact file, however the
+            // kani-metadata file only exists if there were valid targets.
+            trace!(artifact=?path, "map_kani_artifact");
+            let input_file = File::open(path).ok()?;
+            let stub: CompilerArtifactStub = serde_json::from_reader(input_file).unwrap();
+            Artifact::try_new(&stub.metadata_path, ArtifactType::Metadata).ok()
+        }
+    });
+    debug!(?result, "map_kani_artifact");
+    result
+}
+
 /// Possible verification targets.
+#[derive(Debug)]
 enum VerificationTarget {
-    Bin(String),
-    Lib,
-    Test(String),
+    Bin(Target),
+    Lib(Target),
+    Test(Target),
 }
 
 impl VerificationTarget {
     /// Convert to cargo argument that select the specific target.
     fn to_args(&self) -> Vec<String> {
         match self {
-            VerificationTarget::Test(name) => vec![String::from("--test"), name.clone()],
-            VerificationTarget::Bin(name) => vec![String::from("--bin"), name.clone()],
-            VerificationTarget::Lib => vec![String::from("--lib")],
+            VerificationTarget::Test(target) => vec![String::from("--test"), target.name.clone()],
+            VerificationTarget::Bin(target) => vec![String::from("--bin"), target.name.clone()],
+            VerificationTarget::Lib(_) => vec![String::from("--lib")],
+        }
+    }
+
+    fn target(&self) -> &Target {
+        match self {
+            VerificationTarget::Test(target)
+            | VerificationTarget::Bin(target)
+            | VerificationTarget::Lib(target) => target,
+        }
+    }
+}
+
+impl Display for VerificationTarget {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            VerificationTarget::Test(target) => write!(f, "test `{}`", target.name),
+            VerificationTarget::Bin(target) => write!(f, "binary `{}`", target.name),
+            VerificationTarget::Lib(target) => write!(f, "lib `{}`", target.name),
         }
     }
 }
@@ -251,7 +394,7 @@ fn package_targets(args: &KaniArgs, package: &Package) -> Vec<VerificationTarget
             match kind.as_str() {
                 CRATE_TYPE_BIN => {
                     // Binary targets.
-                    verification_targets.push(VerificationTarget::Bin(target.name.clone()));
+                    verification_targets.push(VerificationTarget::Bin(target.clone()));
                 }
                 CRATE_TYPE_LIB | CRATE_TYPE_RLIB | CRATE_TYPE_CDYLIB | CRATE_TYPE_DYLIB
                 | CRATE_TYPE_STATICLIB => {
@@ -264,7 +407,7 @@ fn package_targets(args: &KaniArgs, package: &Package) -> Vec<VerificationTarget
                 CRATE_TYPE_TEST => {
                     // Test target.
                     if args.tests {
-                        verification_targets.push(VerificationTarget::Test(target.name.clone()));
+                        verification_targets.push(VerificationTarget::Test(target.clone()));
                     } else {
                         ignored_tests.push(target.name.as_str());
                     }
@@ -280,7 +423,7 @@ fn package_targets(args: &KaniArgs, package: &Package) -> Vec<VerificationTarget
                         `proc-macro`.",
                 target.name,
             ),
-            (true, false) => verification_targets.push(VerificationTarget::Lib),
+            (true, false) => verification_targets.push(VerificationTarget::Lib(target.clone())),
             (_, _) => {}
         }
     }
